+++ src/Runner.js
@@ -87,17 +87,17 @@ class Runner {
 
 	async deploy() {
 		try {
-			this.log.info('This program will guide you through the deployment of this CloudFlare Worker')
+			this.log.info('This program will guide you through the deployment of this Cloudflare Worker')
 			this.log.info(`Please follow the steps closely. If you want to cancel at anytime, press ${ chalk.cyan('CTRL+C') }`)
 			this.log.line()
 
 			const workerConfig = await parseToml(this.options.template)
 			this.log.debug(workerConfig)
 
-			this.log.load(`Checking if you're authenticated with CloudFlare...`)
+			this.log.load(`Checking if you're authenticated with Cloudflare...`)
 			const isAuthenticated = await wrangler.isAuthenticated()
 			if (!isAuthenticated) {
-				this.log.warn(`Could not authenticate with CloudFlare, you have to login first`)
+				this.log.warn(`Could not authenticate with Cloudflare, you have to login first`)
 				this.log.info(`You can login using your browser or by specifying an API token`)
 
 				const authMethod = await io.selectAuthMethod()
@@ -117,7 +117,7 @@ class Runner {
 				}
 			}
 
-			this.log.succeed(`Successfully authenticated with CloudFlare`)
+			this.log.succeed(`Successfully authenticated with Cloudflare`)
 			this.log.line()
 
 			let accountId = await wrangler.getAccountId()
@@ -237,7 +237,7 @@ class Runner {
 
 			const domainType = await io.selectDomainType()
 			if (domainType === 'Deploy to your own zone') {
-				this.log.info(`Please go to your CloudFlare Dashboard (${ chalk.cyan('https://dash.cloudflare.com') }) to retrieve your ${ chalk.cyan('Zone ID') }`)
+				this.log.info(`Please go to your Cloudflare Dashboard (${ chalk.cyan('https://dash.cloudflare.com') }) to retrieve your ${ chalk.cyan('Zone ID') }`)
 
 				const zoneId = await io.inputZoneId()
 
@@ -289,7 +289,7 @@ class Runner {
 			this.log.line()
 
 			const domain = published.domain ? `https://${ published.domain }` : workerConfig.routes.join(', ')
-			this.log.succeed(`${ chalk.green('Success!') } Your Worker was deployed to ${ chalk.cyan(domain) } 🚀`)
+			this.log.succeed(`${ chalk.green('Success!') } Deployed to ${ chalk.cyan(domain) } 🚀`)
 
 			// Note: Secrets can only be changed after the Worker was pushed i.e. created
 			if (workerConfig.secrets) {
@@ -358,18 +358,23 @@ class Runner {
 				}
 
 				delete workerConfig.secrets
+			}
 
+			this.log.line()
+
+			if (workerConfig.instructions) {
+				this.log.info('Further instructions:')
+				this.log.text(workerConfig.instructions)
 				this.log.line()
 			}
 
 			this.log.succeed('All done! Your Worker is ready to be used ✨')
-			// Print instructions
 			this.log.text(chalk.cyan(domain))
 
 		} catch (err) {
 
 			if (err.name === 'LOGIN') {
-				this.log.fail(`Could not login with CloudFlare. Maybe your token was wrong?`)
+				this.log.fail(`Could not login with Cloudflare. Maybe your token was wrong?`)
 				this.log.info(`Run ${ chalk.cyan('\`wrangler login\`') } or ${ chalk.cyan('\`wrangler config\`') } to login manually or try again.`)
 
 				this.log.debug(err.message)
@@ -464,7 +469,7 @@ class Runner {
 				return
 			}
 
-			this.log.info(`There are multiple ways to login with CloudFlare, using your browser or by specifying an API token`)
+			this.log.info(`There are multiple ways to login with Cloudflare, using your browser or by specifying an API token`)
 
 			const authMethod = await io.selectAuthMethod()
 
@@ -487,7 +492,7 @@ class Runner {
 
 		} catch (err) {
 			if (err.name === 'LOGIN') {
-				this.log.fail(`Could not login with CloudFlare. Maybe your token was wrong?`)
+				this.log.fail(`Could not login with Cloudflare. Maybe your token was wrong?`)
 				this.log.info(`Run ${ chalk.cyan('\`wrangler login\`') } or ${ chalk.cyan('\`wrangler config\`') } to login manually or try again.`)
 
 				this.log.debug(err.message)